+++ apps/health_campaign_field_worker_app/lib/models/request_info/request_info_model.dart
@@ -29,6 +29,7 @@ class RequestInfoModel with _$RequestInfoModel {
     final String? msgId,
     final String? authToken,
     final UserRequestModel? userInfo,
+    final String? tenantId,
   }) = _RequestInfoModel;
 
   factory RequestInfoModel.fromJson(Map<String, dynamic> json) =>