+++ apps/health_campaign_field_worker_app/lib/pages/checklist/checklist_view.dart
@@ -271,8 +271,30 @@ class _ChecklistViewPageState extends LocalizedState<ChecklistViewPage> {
                                             lastModifiedTime: context
                                                 .millisecondsSinceEpoch(),
                                           ),
-                                          additionalDetails:
-                                              context.boundary.code,
+                                          additionalFields:
+                                              ServiceAdditionalFields(
+                                            version: 1,
+                                            fields: [
+                                              AdditionalField(
+                                                'latitude',
+                                                latitude,
+                                              ),
+                                              AdditionalField(
+                                                'longitude',
+                                                longitude,
+                                              ),
+                                              AdditionalField(
+                                                'localityCode',
+                                                context.boundary.code,
+                                              ),
+                                            ],
+                                          ),
+                                          additionalDetails: {
+                                            "boundaryCode":
+                                                context.boundary.code,
+                                            'lat': latitude,
+                                            'lng': longitude,
+                                          },
                                         ),
                                       ),
                                     );